tests/DCPS/UnitTests does not conform to guidelines
Problem
-------

* Unit tests under tests/DCPS/UnitTests do not conform to
  docs/internal/unit_tests.rst.

* tests/DCPS/UnitTests contains an integration test.

* tests/DCPS/UnitTests contains a stress test.

Solution
--------

* Move unit tests to tests/unit-tests.

* Move stress test to tests/stress-tests.

* Move integratin tests to tests/DCPS/BitDataReader.

* Provide a generic run_test.pl for launching unit and stress tests.

* Provide a script for running all stress tests.
