+++ tests/DCPS/BitDataReader/run_test.pl
@@ -1,0 +1,21 @@
+eval '(exit $?0)' && eval 'exec perl -S $0 ${1+"$@"}'
+    & eval 'exec perl -S $0 $argv:q'
+    if 0;
+
+# -*- perl -*-
+
+use Env (DDS_ROOT);
+use lib "$DDS_ROOT/bin";
+use Env (ACE_ROOT);
+use lib "$ACE_ROOT/bin";
+use PerlDDS::Run_Test;
+use strict;
+
+PerlDDS::add_lib_path('../ConsolidatedMessengerIdl');
+
+my $test = new PerlDDS::TestFramework();
+
+$test->process('exec', 'BitDataReader', "-DCPSConfigFile rtps.ini");
+$test->start_process('exec');
+
+exit $test->finish(60);
