--- tests/DCPS/UnitTests/.gitignore
@@ -1,21 +1,0 @@
-/UnitTests_BIT_DataReader
-/UnitTests_DisjointSequence
-/UnitTests_DurationToTimeValue
-/UnitTests_Fragmentation
-/UnitTests_GuidGenerator
-/UnitTests_LivelinessCompatibility
-/UnitTests_MultiTask
-/UnitTests_NetworkConfig
-/UnitTests_ParameterListConverter
-/UnitTests_Reassembly
-/UnitTests_RepoIdSequence
-/UnitTests_RtpsDuration
-/UnitTests_RtpsFragmentation
-/UnitTests_RtpsSequenceNumberSet
-/UnitTests_SequenceNumber
-/UnitTests_TimeTSubtraction
-/UnitTests_XTypesAssignability
-/UnitTests_Serializer
-/UnitTests_GuidUtils
-/UnitTests_JsonValueWriter
-/UnitTests_JsonValueReader
