+++ tests/stress-tests/.gitignore
@@ -1,0 +1,1 @@
+**/StressTests_*
