+++ tests/unit-tests/dds/DCPS/.gitignore
@@ -1,2 +1,1 @@
-/UnitTests_DataSampleHeader
-/SafetyProfilePoolTest
+**/UnitTests_*
