+++ tests/unit-tests/dds/DCPS/GuidUtils.cpp
